fix: move folders outside app
