+++ app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
-import "./styles/globals.css";
-import { ThemeProvider } from "./providers/theme-provider";
+import "@/styles/globals.css";
+import { ThemeProvider } from "@/providers/theme-provider";
 
 export const metadata: Metadata = {
   title: "Steam Shared",
