+++ hooks/use-steam-id-cookie.ts
@@ -1,5 +1,5 @@
 import { useCookies } from "react-cookie";
-import { STEAM_ID_COOKIE } from "../constants";
+import { STEAM_ID_COOKIE } from "@/constants";
 
 export function useSteamIdCookie() {
   const [cookie, setCookies, removeCookies] = useCookies([STEAM_ID_COOKIE]);
