+++ types/page.ts
@@ -1,9 +1,9 @@
-import { FRIEND_ID } from "../constants/search-params";
+import { FRIEND_ID } from "@/constants/search-params";
 
 export type SharedGamesPageSearchParams = {
   [FRIEND_ID]: string;
 };
 
 export type SharedGamesPageProps = {
-  searchParams: SharedGamesPageSearchParams;
+  searchParams: Promise<SharedGamesPageSearchParams>;
 };
